Homework 3
Read some string str.Calculate the counts of characters ‘a’, ’o’, ’i’, ’e’  in this text.
Ask user to enter the number of month. Read value and output the count of days in this month.
Enter 10 integer numbers. Calculate the sum of first 5 elements if they are positive or product of last 5 element in  the other case.